refactor: ReservationTime 필드 final 추가, getter 수정, 패키지 이동
