+++ src/main/java/roomescape/domain/ReservationTime.java
@@ -1,0 +1,21 @@
+package roomescape.domain;
+
+import java.time.LocalTime;
+
+public class ReservationTime {
+    private final Long id;
+    private final LocalTime startAt;
+
+    public ReservationTime(Long id, LocalTime startAt) {
+        this.id = id;
+        this.startAt = startAt;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public LocalTime getStartAt() {
+        return startAt;
+    }
+}
